vector_selection.cc: Extract vector_selection_(internal|take|filter).(cc|h)
